structuring the code
